update menu modal & refactor code
