+++ src/components/modals/MenuModal/MenuModal.css
@@ -1,0 +1,3 @@
+.menu-modal .input-wrapper {
+  margin-bottom: 10px;
+}
