+++ src/components/modals/MenuModal/MenuModal.tsx
@@ -1,11 +1,13 @@
-import React, { useEffect } from "react";
+import React, { useEffect, useState } from "react";
 import { Modal, Form, Input, InputNumber, Select, Switch } from "antd";
-import { MenuItem } from "../interfaces/Menu.interface";
+
+import { MenuItem } from "../../../interfaces/Menu.interface";
+import "./MenuModal.css";
 
 interface MenuModalProps {
   visible: boolean;
   menu?: MenuItem;
-  onSave: (menu: MenuItem) => void;
+  onSave: (menu: MenuItem) => Promise<void | undefined>;
   onCancel: () => void;
   categories: { id: string; name: string }[];
 }
@@ -18,17 +20,11 @@ const MenuModal: React.FC<MenuModalProps> = ({
   categories,
 }) => {
   const [form] = Form.useForm();
-
-  useEffect(() => {
-    if (menu) {
-      form.setFieldsValue(menu);
-    } else {
-      form.resetFields();
-    }
-  }, [menu]);
+  const [isLoading, setIsLoading] = useState(false);
 
   const handleSubmit = () => {
     form.validateFields().then((values) => {
+      setIsLoading(true);
       // Find the category name based on the selected category_id
       const selectedCategory = categories.find(
         (category) => category.id === values.category_id
@@ -36,35 +32,65 @@ const MenuModal: React.FC<MenuModalProps> = ({
       const categoryName = selectedCategory ? selectedCategory.name : "";
 
       // Include both category_id and category_name in the submitted values
-      onSave({ ...menu, ...values, category: categoryName });
+      onSave({ ...menu, ...values, category: categoryName }).finally(() =>
+        setIsLoading(false)
+      );
     });
   };
 
+  useEffect(() => {
+    if (menu) {
+      form.setFieldsValue(menu);
+    } else {
+      form.resetFields();
+    }
+  }, [menu, form]);
+
   return (
     <Modal
-      open={visible}
-      onCancel={onCancel}
-      onOk={handleSubmit}
       title={menu ? "Edit Menu" : "Add Menu"}
+      open={visible}
+      centered={true}
+      confirmLoading={isLoading}
+      onCancel={onCancel}
+      okText={menu ? "Save Changes" : "Add"}
+      onOk={handleSubmit}
     >
-      <Form form={form} layout="vertical">
-        <Form.Item name="name" label="Name" rules={[{ required: true }]}>
+      <Form form={form} layout="vertical" className="menu-modal">
+        <Form.Item
+          className="input-wrapper"
+          name="name"
+          label="Name"
+          rules={[{ required: true }]}
+        >
           <Input />
         </Form.Item>
         <Form.Item
           name="description"
+          className="input-wrapper"
           label="Description"
           rules={[{ required: true }]}
         >
           <Input />
         </Form.Item>
-        <Form.Item name="price" label="Price" rules={[{ required: true }]}>
+        <Form.Item
+          className="input-wrapper"
+          name="price"
+          label="Price"
+          rules={[{ required: true }]}
+        >
           <InputNumber min={0} style={{ width: "100%" }} />
         </Form.Item>
-        <Form.Item name="rating" label="Rating" rules={[{ required: true }]}>
+        <Form.Item
+          className="input-wrapper"
+          name="rating"
+          label="Rating"
+          rules={[{ required: true }]}
+        >
           <InputNumber min={1} max={5} />
         </Form.Item>
         <Form.Item
+          className="input-wrapper"
           name="quantity"
           label="Quantity"
           rules={[{ required: true }]}
@@ -72,6 +98,7 @@ const MenuModal: React.FC<MenuModalProps> = ({
           <InputNumber />
         </Form.Item>
         <Form.Item
+          className="input-wrapper"
           name="category_id"
           label="Category"
           rules={[{ required: true }]}
@@ -85,6 +112,7 @@ const MenuModal: React.FC<MenuModalProps> = ({
           </Select>
         </Form.Item>
         <Form.Item
+          className="input-wrapper"
           name="is_available"
           label="Available"
           valuePropName="checked"
